Add working fillInterp that uses boxes
Unlike older commits, this still looks up the stencil for every ghost
cell when required rather than using the same stencil for the whole box
of ghosts. The main advantage of boxes is the ability to move the OpenMP
loop over the box rather than the components which is done here.